Fix Timeout narrowing in CprHttpClient

+++ src/core/net/cpr_http_client.cpp
@@ -9,7 +9,7 @@ HttpResponse CprHttpClient::get(const std::string &url,
                                 const std::map<std::string, std::string> &headers) {
   HttpResponse resp;
   try {
-    cpr::Timeout to{timeout.count()};
+    cpr::Timeout to{static_cast<int32_t>(timeout.count())};
     cpr::Header hdr{headers.begin(), headers.end()};
     auto r = cpr::Get(cpr::Url{url}, to, hdr);
     resp.status_code = static_cast<int>(r.status_code);
